convert enqueue_job into a method of request_handler
