+++ x-pack/plugins/reporting/server/routes/lib/request_handler.ts
@@ -10,8 +10,8 @@ import { KibanaRequest, KibanaResponseFactory } from 'kibana/server';
 import { ReportingCore } from '../..';
 import { API_BASE_URL } from '../../../common/constants';
 import { JobParamsPDFLegacy } from '../../export_types/printable_pdf/types';
-import { LevelLogger } from '../../lib';
-import { enqueueJob } from '../../lib/enqueue_job';
+import { checkParamsVersion, cryptoFactory, LevelLogger } from '../../lib';
+import { Report } from '../../lib/store';
 import { BaseParams, ReportingRequestHandlerContext, ReportingUser } from '../../types';
 
 export const handleUnavailable = (res: KibanaResponseFactory) => {
@@ -32,6 +32,72 @@ export class RequestHandler {
     private res: KibanaResponseFactory,
     private logger: LevelLogger
   ) {}
+
+  private async encryptHeaders() {
+    const encryptionKey = this.reporting.getConfig().get('encryptionKey');
+    const crypto = cryptoFactory(encryptionKey);
+    return await crypto.encrypt(this.req.headers);
+  }
+
+  public async enqueueJob(exportTypeId: string, jobParams: BaseParams) {
+    const { reporting, logger, context, req: request, user } = this;
+
+    const exportType = reporting.getExportTypesRegistry().getById(exportTypeId);
+
+    if (exportType == null) {
+      throw new Error(`Export type ${exportTypeId} does not exist in the registry!`);
+    }
+
+    if (!exportType.createJobFnFactory) {
+      throw new Error(`Export type ${exportTypeId} is not an async job type!`);
+    }
+
+    const [createJob, store] = await Promise.all([
+      exportType.createJobFnFactory(reporting, logger.clone([exportType.id])),
+      reporting.getStore(),
+    ]);
+
+    if (!createJob) {
+      throw new Error(`Export type ${exportTypeId} is not an async job type!`);
+    }
+
+    // 1. ensure the incoming params have a version field
+    jobParams.version = checkParamsVersion(jobParams, logger);
+
+    // 2. encrypt request headers for the running report job to authenticate itself with Kibana
+    // 3. create the payload with the queued job info
+    // 4. get the Space ID from the request
+    const [job, headers, spaceId] = await Promise.all([
+      createJob(jobParams, context),
+      this.encryptHeaders(),
+      reporting.getSpaceId(request, logger),
+    ]);
+    const payload = { ...job, headers, spaceId };
+
+    // 5. Add the report to ReportingStore to show as pending
+    const report = await store.addReport(
+      new Report({
+        jobtype: exportType.jobType,
+        created_by: user ? user.username : false,
+        payload,
+        meta: {
+          // telemetry fields
+          objectType: jobParams.objectType,
+          layout: jobParams.layout?.id,
+          isDeprecated: job.isDeprecated,
+        },
+      })
+    );
+    logger.debug(`Successfully stored pending job: ${report._index}/${report._id}`);
+
+    // 6. Schedule the report with Task Manager
+    const task = await reporting.scheduleTask(report.toReportTaskJSON());
+    logger.info(
+      `Scheduled ${exportType.name} reporting task. Task ID: task:${task.id}. Report ID: ${report._id}`
+    );
+
+    return report;
+  }
 
   public async handleGenerateRequest(
     exportTypeId: string,
@@ -54,15 +120,7 @@ export class RequestHandler {
     }
 
     try {
-      const report = await enqueueJob(
-        this.reporting,
-        this.req,
-        this.context,
-        this.user,
-        exportTypeId,
-        jobParams,
-        this.logger
-      );
+      const report = await this.enqueueJob(exportTypeId, jobParams);
 
       // return task manager's task information and the download URL
       const downloadBaseUrl = getDownloadBaseUrl(this.reporting);
